fix: array.hh  / postprocessing.cc fix initialize type warnings

+++ Holovibes/includes/tools/cuda_tools/array.hh
@@ -95,7 +95,7 @@ namespace holovibes
 			std::vector<T> to_cpu()
 			{
 				std::vector<T> cpu_buffer(size_);
-				const uint byte_size = size_ * sizeof(T);
+				const size_t byte_size = size_ * sizeof(T);
 				cudaMemcpy(cpu_buffer.data(), get(), byte_size, cudaMemcpyDeviceToHost);
 				return cpu_buffer;
 			}
